Refactor AST and Parser separation
- Extracted parsing logic from `src/ast/mod.rs` to `src/parser/builder.rs`.
- Created `src/parser/mod.rs` as a facade for the parser.
- Renamed `AstError` to `ParseError` and moved it to `src/parser/builder.rs`.
- Updated all consumers (main, tests, semantic, codegen, ir) to use `glossa::parser::parse`.
- `src/ast` is now a pure data module containing only node definitions.
- Fixed a doctest in `src/ir/hir.rs` that was referencing the old `build_ast` function.

This change enforces better architectural boundaries by decoupling the AST data structure from its construction logic.
