+++ tests/compile_tests.rs
@@ -1,13 +1,13 @@
 //! Integration tests for the ΓΛΩΣΣΑ compiler pipeline
 
-use glossa::ast::build_ast;
 use glossa::codegen::generate_rust;
 use glossa::ir::lower_to_hir;
+use glossa::parser::parse;
 use glossa::semantic::analyze_program;
 
 /// Helper to compile GLOSSA source to Rust code
 fn compile(source: &str) -> Result<String, String> {
-    let ast = build_ast(source).map_err(|e| e.to_string())?;
+    let ast = parse(source).map_err(|e| e.to_string())?;
     let analyzed = analyze_program(&ast).map_err(|e| e.to_string())?;
     let hir = lower_to_hir(&analyzed);
     Ok(generate_rust(&hir))
