+++ tests/control_flow_tests.rs
@@ -1,3 +1,5 @@
+use glossa::codegen::generate_rust;
+use glossa::ir::lower_to_hir;
 /// Phase 2: Control Flow Tests
 /// TDD RED phase - these tests define the expected behavior
 ///
@@ -6,13 +8,11 @@
 /// - Loops: ἕως (while), διά/ἀπό...μέχρι (for)
 /// - Pattern matching: κατά (match)
 /// - Loop control: παῦε (break), συνέχιζε (continue)
-use glossa::ast::build_ast;
-use glossa::codegen::generate_rust;
-use glossa::ir::lower_to_hir;
+use glossa::parser::parse;
 use glossa::semantic::analyze_program;
 
 fn compile_to_rust(source: &str) -> String {
-    let ast = build_ast(source).expect("AST build failed");
+    let ast = parse(source).expect("AST build failed");
     let analyzed = analyze_program(&ast).expect("Analysis failed");
     let hir = lower_to_hir(&analyzed);
     generate_rust(&hir)
